server/rag_service/requirements.txt
# server/rag_service/requirements.txt
flask
requests
# faiss-cpu # or faiss-gpu
langchain
langchain-huggingface
pypdf
PyPDF2
python-docx
python-dotenv
ollama # Keep if using Ollama embeddings
python-pptx
uuid
langchain-community
pdfplumber
fitz # PyMuPDF for PDF parsing
pytesseract
nltk
spacy-layout
pandas
numpy
typing
pytesseract # OCR
pillow
qdrant-client
neo4j
sentence_transformers
spacy
opencv-python
ddgs
python-pptx
python-docx
reportlab
google-generativeai
gTTS
yt-dlp
openai-whisper
beautifulsoup4
playwright
moviepy